refactor: move presets to lib/config
