+++ lib/config/presets/__fixtures__/renovate-config-default.json
