+++ lib/config/presets/__fixtures__/renovate-config-group.json
