+++ lib/config/presets/__fixtures__/renovate-config-ikatyang.json
