+++ lib/config/presets/__fixtures__/renovate-config-monorepo.json
