+++ lib/config/presets/__fixtures__/renovate-config-packages.json
