+++ lib/config/presets/index.ts
@@ -2,9 +2,9 @@ import is from '@sindresorhus/is';
 import { logger } from '../../logger';
 import * as massage from '../massage';
 import * as migration from '../migration';
-import * as github from '../../datasource/github';
-import * as npm from '../../datasource/npm';
-import * as gitlab from '../../datasource/gitlab';
+import * as github from './github';
+import * as npm from './npm';
+import * as gitlab from './gitlab';
 import { RenovateConfig } from '../common';
 import { mergeChildConfig } from '../utils';
 import { regEx } from '../../util/regex';
