+++ lib/datasource/github/index.spec.ts
@@ -1,15 +1,12 @@
 import { api } from '../../platform/github/gh-got-wrapper';
 
 import * as github from '.';
-import _got from '../../util/got';
 import * as _hostRules from '../../util/host-rules';
-import { PLATFORM_FAILURE } from '../../constants/error-messages';
 
 jest.mock('../../platform/github/gh-got-wrapper');
 jest.mock('../../util/got');
 jest.mock('../../util/host-rules');
 
-const got: any = _got;
 const ghGot: any = api.get;
 const hostRules: any = _hostRules;
 
@@ -60,61 +57,6 @@ describe('datasource/github', () => {
       ghGot.mockReturnValueOnce({});
       const res = await github.getDigest({ lookupName: 'some/dep' }, 'v1.2.0');
       expect(res).toBeNull();
-    });
-  });
-  describe('getPreset()', () => {
-    it('passes up platform-failure', async () => {
-      got.mockImplementationOnce(() => {
-        throw new Error(PLATFORM_FAILURE);
-      });
-      await expect(github.getPreset('some/repo')).rejects.toThrow(
-        PLATFORM_FAILURE
-      );
-    });
-    it('tries default then renovate', async () => {
-      got.mockImplementationOnce(() => {
-        throw new Error();
-      });
-      await expect(github.getPreset('some/repo')).rejects.toThrow();
-    });
-    it('throws if no content', async () => {
-      got.mockImplementationOnce(() => ({
-        body: {},
-      }));
-      await expect(github.getPreset('some/repo')).rejects.toThrow();
-    });
-    it('throws if fails to parse', async () => {
-      got.mockImplementationOnce(() => ({
-        body: {
-          content: Buffer.from('not json').toString('base64'),
-        },
-      }));
-      await expect(github.getPreset('some/repo')).rejects.toThrow();
-    });
-    it('should return default.json', async () => {
-      hostRules.find.mockReturnValueOnce({ token: 'abc' });
-      got.mockImplementationOnce(() => ({
-        body: {
-          content: Buffer.from('{"foo":"bar"}').toString('base64'),
-        },
-      }));
-      const content = await github.getPreset('some/repo');
-      expect(content).toEqual({ foo: 'bar' });
-    });
-    it('should return custom.json', async () => {
-      hostRules.find.mockReturnValueOnce({ token: 'abc' });
-      got.mockImplementationOnce(() => ({
-        body: {
-          content: Buffer.from('{"foo":"bar"}').toString('base64'),
-        },
-      }));
-      try {
-        global.appMode = true;
-        const content = await github.getPreset('some/repo', 'custom');
-        expect(content).toEqual({ foo: 'bar' });
-      } finally {
-        delete global.appMode;
-      }
     });
   });
   describe('getPkgReleases', () => {
