+++ lib/datasource/gitlab/index.spec.ts
@@ -11,44 +11,6 @@ describe('datasource/gitlab', () => {
     global.repoCache = {};
     return global.renovateCache.rmAll();
   });
-  describe('getPreset()', () => {
-    it('throws if non-default', async () => {
-      await expect(
-        gitlab.getPreset('some/repo', 'non-default')
-      ).rejects.toThrow();
-    });
-    it('throws if no content', async () => {
-      glGot.mockImplementationOnce(() => ({
-        body: {},
-      }));
-      await expect(gitlab.getPreset('some/repo')).rejects.toThrow();
-    });
-    it('throws if fails to parse', async () => {
-      glGot.mockImplementationOnce(() => ({
-        body: {
-          content: Buffer.from('not json').toString('base64'),
-        },
-      }));
-      await expect(gitlab.getPreset('some/repo')).rejects.toThrow();
-    });
-    it('should return the preset', async () => {
-      glGot.mockResolvedValueOnce({
-        body: [
-          {
-            name: 'master',
-            default: true,
-          },
-        ],
-      });
-      glGot.mockResolvedValueOnce({
-        body: {
-          content: Buffer.from('{"foo":"bar"}').toString('base64'),
-        },
-      });
-      const content = await gitlab.getPreset('some/repo');
-      expect(content).toEqual({ foo: 'bar' });
-    });
-  });
   describe('getPkgReleases', () => {
     beforeAll(() => global.renovateCache.rmAll());
     it('returns releases', async () => {
