+++ lib/datasource/npm/index.spec.ts
@@ -74,108 +74,6 @@ describe('api/npm', () => {
     expect(getRelease(res, '0.0.1').canBeUnpublished).toBe(false);
     expect(getRelease(res, '0.0.2').canBeUnpublished).toBe(false);
   });
-  it('should throw if no package', async () => {
-    nock('https://registry.npmjs.org')
-      .get('/foobar')
-      .reply(404);
-    await expect(npm.getPreset('foobar', 'default')).rejects.toThrow(
-      /dep not found/
-    );
-  });
-  it('should throw if no renovate-config', async () => {
-    const presetPackage = {
-      name: 'foobar',
-      versions: {
-        '0.0.1': {
-          foo: 1,
-        },
-        '0.0.2': {
-          foo: 2,
-          deprecated: 'This is deprecated',
-        },
-      },
-      repository: {
-        type: 'git',
-        url: 'git://github.com/renovateapp/dummy.git',
-      },
-      'dist-tags': {
-        latest: '0.0.2',
-      },
-      time: {
-        '0.0.1': '2018-05-06T07:21:53+02:00',
-        '0.0.2': '2018-05-07T07:21:53+02:00',
-      },
-    };
-    nock('https://registry.npmjs.org')
-      .get('/foobar')
-      .reply(200, presetPackage);
-    await expect(npm.getPreset('foobar', 'default')).rejects.toThrow(
-      /preset renovate-config not found/
-    );
-  });
-  it('should throw if no preset name not found', async () => {
-    const presetPackage = {
-      name: 'foobar',
-      versions: {
-        '0.0.1': {
-          foo: 1,
-        },
-        '0.0.2': {
-          foo: 2,
-          deprecated: 'This is deprecated',
-          'renovate-config': { default: { rangeStrategy: 'auto' } },
-        },
-      },
-      repository: {
-        type: 'git',
-        url: 'git://github.com/renovateapp/dummy.git',
-      },
-      'dist-tags': {
-        latest: '0.0.2',
-      },
-      time: {
-        '0.0.1': '2018-05-06T07:21:53+02:00',
-        '0.0.2': '2018-05-07T07:21:53+02:00',
-      },
-    };
-    nock('https://registry.npmjs.org')
-      .get('/foobar')
-      .reply(200, presetPackage);
-    await expect(npm.getPreset('foobar', 'missing')).rejects.toThrow(
-      /preset not found/
-    );
-  });
-  it('should return preset', async () => {
-    const presetPackage = {
-      name: 'foobar',
-      versions: {
-        '0.0.1': {
-          foo: 1,
-        },
-        '0.0.2': {
-          foo: 2,
-          deprecated: 'This is deprecated',
-          'renovate-config': { default: { rangeStrategy: 'auto' } },
-        },
-      },
-      repository: {
-        type: 'git',
-        url: 'https://github.com/renovateapp/dummy.git',
-      },
-      'dist-tags': {
-        latest: '0.0.2',
-      },
-      time: {
-        '0.0.1': '2018-05-06T07:21:53+02:00',
-        '0.0.2': '2018-05-07T07:21:53+02:00',
-      },
-    };
-    nock('https://registry.npmjs.org')
-      .get('/foobar')
-      .reply(200, presetPackage);
-    const res = await npm.getPreset('foobar');
-    expect(res).toMatchSnapshot();
-  });
   it('should parse repo url', async () => {
     const pkg = {
       name: 'foobar',
