+++ lib/datasource/npm/index.ts
@@ -1,4 +1,3 @@
 export { resetMemCache, resetCache } from './get';
-export { getPreset } from './presets';
 export { getPkgReleases } from './releases';
 export { setNpmrc } from './npmrc';
