+++ test/config/index.spec.ts
@@ -1,7 +1,6 @@
 import getArgv from './config/_fixtures/argv';
 import { getConfig } from '../../lib/config/defaults';
 import * as _npm from '../../lib/datasource/npm';
-import presetDefaults from './npm/_fixtures/renovate-config-default.json';
 
 jest.mock('../../lib/datasource/npm');
 try {
@@ -10,15 +9,7 @@ try {
   // file does not exist
 }
 
-const npm: any = _npm;
 const defaultConfig = getConfig();
-
-npm.getPkgReleases = jest.fn(() => ({
-  'renovate-config':
-    presetDefaults.versions[presetDefaults['dist-tags'].latest][
-      'renovate-config'
-    ],
-}));
 
 describe('config/index', () => {
   describe('.parseConfigs(env, defaultArgv)', () => {
